refactor(7092): isolate live settlement orchestration
Move live intent preparation, submission, evidence validation, and typed responses into focused route modules while retaining shared authorization and legacy slot-backed behavior in the parent.\n\nConstraint: Files must remain under 200 lines and functions under 25 lines\nConfidence: high\nScope-risk: narrow\nReversibility: clean\nTested: Settlement-focused contracts (8 passed); strict kamn-node clippy\nNot-tested: Concurrent release and full node suite
